+++ resources/assets/js/dashboard/components/user-createedit.js
@@ -45,6 +45,10 @@ export default {
         },
 
         destroy() {
+            if (!confirm(this.$trans('Do you really want to delete it?'))) {
+                return;
+            }
+
             axios.delete(Laravel.url(`/dashboard/api/users/${this.modelId}`))
                 .then(response => {
                     window.location = Laravel.url('/dashboard/users');
